+++ src/actions/order/new.test.ts
@@ -11,13 +11,14 @@ import { describe, expect, inject, it } from 'vitest'
 import { limitOrderResultFromLogs } from '~mgv/lib/limit-order.js'
 import { tickFromVolumes } from '~mgv/lib/tick.js'
 import { getClient } from '~test/src/client.js'
+import { getMarkets } from '~test/src/markets.js'
 import { BS, Order } from '../../lib/enums.js'
 import { getBook } from '../book.js'
 import { getUserRouter } from '../smart-router.js'
 import { getLimitOrderSteps, simulateLimitOrder } from './new.js'
 
 const params = inject('mangrove')
-const { wethUSDC } = inject('markets')
+const { wethUSDC } = getMarkets()
 const client = getClient()
 
 describe('new order', () => {