+++ src/addresses/tokens/index.ts
@@ -16,6 +16,15 @@ export {
   baseSepoliaTokens,
 } from './base-sepolia.js'
 
+export {
+  arbitrumWETH,
+  arbitrumWBTC,
+  arbitrumUSDC,
+  arbitrumUSDT,
+  arbitrumweETH,
+  arbitrumTokens,
+} from './arbitrum.js'
+
 export type {
   Token,
   BuildTokenParms,