+++ src/index.ts
@@ -87,6 +87,7 @@ export {
 export type {
   MangroveActionsDefaultParams,
   MarketParams,
+  SerializableMarketParams,
   BuiltArgs,
   BuiltArgsWithValue,
   Book,
@@ -105,6 +106,7 @@ export type {
   LocalConfig,
   CompleteToken,
   OLKey,
+  SerializableOLKey,
   RpcOffer,
   RpcOfferDetail,
   RpcCompleteOffer,
@@ -120,6 +122,7 @@ export type {
   KandelActions,
   KandelSeederActions,
   UserRouterActions,
+  AaveKandelActions,
 } from './bundle/index.js'
 
 export {
@@ -129,6 +132,7 @@ export {
   kandelActions,
   kandelSeederActions,
   userRouterActions,
+  aaveKandelActions,
 } from './bundle/index.js'
 
 // --- addresses ---